Day 5
A bit tedious to parse, but removing from a queue ended up making it
easy.

Expanding the ranges in the map was not possible because you ran out of
memory, but given there were only few ranges, iterating over the list to
find the matching range was not too bad.

In the second part, the challenge was processing the massive ranges of
seeds without keeping the results in memory, and not expanding the
ranges and folding to keep just the current minimum should do that. In
the end, it took a while to finish, but got the right answered.