Day 7
Kind of hard to share the code for both parts. In the end, just
extracted the way hands get compared to be passed in rather than making
the hand and card comparables.

As I'm not using packages, I had to rename a class from day 4 so that
there was no name clash. Need to find a proper solution for it later.