Day 8
The pure recursive option didn't work for the second part, even though I
think it was correct, but it took too long, it never finished no matter
how long I waited, and you can tell why when you see the result reaching
10241191004509.

In the end, the issue was that it required to traverse the instruction
so many times before all the start nodes reach the end node at the same
time that it wasn't feasible to do the full traversal. Instead, you
could find how many steps it took to find the end node for each start
node, and find the least common multiple of all the results to find the
shortest path without actually going through it.