Day 9
Most enjoyable so far. Nice to play with recursion and trees.

Suffixed classes with History to avoid clashing with other classes.
Tried to add solutions to packages, but liked the simplicity of just
having files in the root, so decided that giving more specific name to
things would be a good exercise.