Day 11
Initially expanded the universe by actually adding columns and rows to
the 2d list by transposing it so that I could treat columns as rows.
This didn't work for the 2nd part, as the universe grew too much that it
didn't fit in memory, but got there in the end by calculating the new
position of the galaxies.